Complete hardening follow-ups for OpenAPI, CI, and startup resilience
