+++ .github/workflows/ci.yml
@@ -112,7 +112,7 @@ jobs:
       - name: Enforce minimum coverage
         shell: bash
         env:
-          MIN_COVERAGE: "40.5"
+          MIN_COVERAGE: "40.6"
         run: |
           set -euo pipefail
           mapfile -t all_packages < <(go list ./...)
@@ -137,13 +137,13 @@ jobs:
           from pathlib import Path
 
           required = {
-              "github.com/writer/cerebro/internal/api": 35.0,
-              "github.com/writer/cerebro/internal/app": 45.0,
-              "github.com/writer/cerebro/internal/cache": 80.0,
-              "github.com/writer/cerebro/internal/graph": 60.0,
-              "github.com/writer/cerebro/internal/policy": 70.0,
-              "github.com/writer/cerebro/internal/scanner": 35.0,
-              "github.com/writer/cerebro/internal/scheduler": 70.0,
+              "github.com/writer/cerebro/internal/api": 36.0,
+              "github.com/writer/cerebro/internal/app": 47.0,
+              "github.com/writer/cerebro/internal/cache": 82.0,
+              "github.com/writer/cerebro/internal/graph": 62.0,
+              "github.com/writer/cerebro/internal/policy": 72.0,
+              "github.com/writer/cerebro/internal/scanner": 36.0,
+              "github.com/writer/cerebro/internal/scheduler": 71.5,
           }
 
           pattern = re.compile(r"^ok\s+(\S+)\s+.*coverage:\s+([0-9.]+)% of statements$")
@@ -204,6 +204,109 @@ jobs:
 
       - name: Verify OpenAPI matches registered routes
         run: go run ./scripts/openapi_route_parity.go
+
+      - name: Ensure generated OpenAPI placeholders are removed
+        shell: bash
+        run: |
+          set -euo pipefail
+          if grep -n "x-cerebro-generated" api/openapi.yaml; then
+            echo "OpenAPI still contains generated placeholders. Replace with real operation docs."
+            exit 1
+          fi
+
+  config-docs-drift:
+    name: config-docs-drift
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: '1.25.7'
+
+      - name: Regenerate config docs
+        run: go run ./scripts/generate_config_docs/main.go
+
+      - name: Verify generated config docs are committed
+        run: git diff --exit-code -- docs/CONFIG_ENV_VARS.md
+
+  api-auth-smoke:
+    name: api-auth-smoke
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: '1.25.7'
+
+      - name: Build
+        run: go build -o cerebro ./cmd/cerebro
+
+      - name: API auth end-to-end smoke
+        shell: bash
+        env:
+          LOG_LEVEL: error
+          API_PORT: "18080"
+          API_AUTH_ENABLED: "true"
+          API_KEYS: "smoke-key:smoke-admin"
+          RBAC_STATE_FILE: "/tmp/cerebro-rbac-smoke.json"
+        run: |
+          set -euo pipefail
+          cat > "$RBAC_STATE_FILE" <<'JSON'
+          {
+            "users": {
+              "smoke-admin": {
+                "id": "smoke-admin",
+                "email": "smoke-admin@example.com",
+                "name": "Smoke Admin",
+                "role_ids": ["admin"]
+              }
+            }
+          }
+          JSON
+
+          ./cerebro serve > /tmp/cerebro-api-auth-smoke.log 2>&1 &
+          server_pid=$!
+          cleanup() {
+            kill "$server_pid" >/dev/null 2>&1 || true
+            wait "$server_pid" >/dev/null 2>&1 || true
+          }
+          trap cleanup EXIT
+
+          for i in $(seq 1 45); do
+            if curl -fsS "http://127.0.0.1:${API_PORT}/health" >/dev/null 2>&1; then
+              break
+            fi
+            sleep 1
+          done
+          curl -fsS "http://127.0.0.1:${API_PORT}/health" >/dev/null
+
+          unauth_status=$(curl -sS -o /tmp/api-auth-smoke-unauth.json -w "%{http_code}" \
+            "http://127.0.0.1:${API_PORT}/api/v1/policies/")
+          if [ "$unauth_status" != "401" ]; then
+            echo "expected 401 for unauthenticated request, got $unauth_status"
+            cat /tmp/cerebro-api-auth-smoke.log || true
+            exit 1
+          fi
+
+          auth_status=$(curl -sS -H "Authorization: Bearer smoke-key" -o /tmp/api-auth-smoke-auth.json -w "%{http_code}" \
+            "http://127.0.0.1:${API_PORT}/api/v1/policies/")
+          if [ "$auth_status" -ne "200" ]; then
+            echo "expected 200 for authenticated request, got $auth_status"
+            cat /tmp/cerebro-api-auth-smoke.log || true
+            exit 1
+          fi
+
+          xkey_status=$(curl -sS -H "X-API-Key: smoke-key" -o /tmp/api-auth-smoke-xkey.json -w "%{http_code}" \
+            "http://127.0.0.1:${API_PORT}/api/v1/policies/")
+          if [ "$xkey_status" -ne "200" ]; then
+            echo "expected 200 for X-API-Key request, got $xkey_status"
+            cat /tmp/cerebro-api-auth-smoke.log || true
+            exit 1
+          fi
 
   scan-smoke-local:
     name: scan-smoke-local
