+++ scripts/doc.go
@@ -1,0 +1,2 @@
+// Package scripts hosts repository maintenance tools that are executed via `go run`.
+package scripts
