+++ scripts/openapi_route_parity.go
@@ -1,3 +1,5 @@
+//go:build ignore
+
 package main
 
 import (
